Remove the dam profile table and render the dam svg in the DamProfile component, instead of the DamProfileChart component. Rendering the svg directly in the DamProfile component removes the need to do useMemo
